+++ index.html
@@ -84,12 +84,15 @@
                 </select>
               </div>
               <button id="create" class="create__note-btn">Save</button>
-              <!-- <input id="edit" type="button" value="Change" class="create__note-btn"> -->
             </form>
           </div>
           <div id="modalArchive" class="notes__list hide"></div>
       </div>
     </main>
-    <script type="module" src="/Script/index.js"></script>
+    <script type="module" src="/src/js/modules/DataModule.js"></script>
+    <script type="module" src="/src/js/modules/DOMModule.js"></script>
+    <script type="module" src="/src/js/modules/ModalModule.js"></script>
+    <script type="module" src="/src/js/modules/NotesModule.js"></script>
+    <script type="module" src="/src/js/App.js"></script>
   </body>
 </html>
